Add files via upload
The ultimate goal is to find the ratings given by the user to the product using the Collaborative filtering method involves Singular Value Decomposition (SVD) and Stochastic Gradient Descent (SGD). In this paper, we find the evaluation between SVD and SGD, in order to determine which approach gives the best accuracy of the rating.